adapted to number 22

+++ EvalFive.py
@@ -373,11 +373,11 @@ def test():
 
 	print '-------- EvalFive getFiveRank_ getSevenRank_ evaluators sample runs'
 	np.random.seed(99)
-	array_five_cards = np.array([np.random.choice(keys.DECK_SIZE, 5, replace=False) for k in xrange(10)])
+	array_five_cards = np.array([np.random.choice(keys.DECK_SIZE, 5, replace=False) for k in xrange(10)], dtype=np.int32)
 	print 'array_five_cards=\n{}'.format(array_five_cards)
 	five_cards = array_five_cards[0]
 	print 'five_cards={}'.format(five_cards)
-	array_seven_cards = np.array([np.random.choice(keys.DECK_SIZE, 7, replace=False) for k in xrange(10)])
+	array_seven_cards = np.array([np.random.choice(keys.DECK_SIZE, 7, replace=False) for k in xrange(10)], dtype=np.int32)
 	print 'array_seven_cards={}\n'.format(array_seven_cards)
 	seven_cards = array_seven_cards[0]
 	print 'seven_cards={}'.format(seven_cards)
